+++ src/assets/fuelcell.jl
@@ -7,7 +7,7 @@ mutable struct FuelCell
      α_p::Float64
      η_H2_E::Float64
      η_H2_H::Float64
-     lifetime::Float64
+     lifetime::Int64
      nHoursMax::Float64
      # Initial conditions
      powerMax_ini::Float64
